better debug routines, code cleaning
Added a couple more debug groups to break up the amount of info
presented into clearer groups. Cleaned up some old code that was
commented out and no longer in use.

+++ motorControler.ino
@@ -12,7 +12,7 @@ void initMotorControler()
   Serial.println("Motor controler: configured and active");
 }
 
-void motorControl1(long value)
+void motorControl1(float value)
 {
   value = constrain(value, -255, 255);
    
@@ -57,7 +57,7 @@ void motorControl1(long value)
   
 }
 
-void motorControl2(long value)
+void motorControl2(float value)
 {
   
   if(value < 0)
